openclaw-skills/requirements.txt
httpx>=0.27,<1.0
discord.py>=2.3,<3.0
python-telegram-bot>=21.0,<22.0
pydantic>=2.5,<3.0
python-dotenv>=1.0,<2.0
anthropic>=0.40,<1.0
